# packages/contracts/openapi/booking.v1.yaml
# Copyright (c) 2026, SoftlaneIT (https://softlaneit.com/) All Rights Reserved.
#
# SoftlaneIT licenses this file to you under the Apache License,
# Version 2.0 (the "LICENSE"); you may not use this file except
# in compliance with the LICENSE.
# You may obtain a copy of the LICENSE at
#
# https://softlaneit.com/LICENSE.txt
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the LICENSE is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the LICENSE for the
# specific language governing permissions and limitations
# under the LICENSE.

openapi: 3.0.3
info:
  title: ServiceForge Booking API
  version: 1.0.0
servers:
  - url: http://localhost:8085
paths:
  /v1/bookings:
    get:
      summary: List bookings for tenant
      parameters:
        - in: header
          name: X-Tenant-ID
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Booking list
    post:
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customer, service, startAt]
              properties:
                customer:
                  type: string
                service:
                  type: string
                startAt:
                  type: string
                  format: date-time
      responses:
        "201":
          description: Booking created
  /v1/bookings/{bookingId}:
    get:
      summary: Get a booking by id
      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking details
    put:
      summary: Update booking status
      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking updated
    delete:
      summary: Delete booking
      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking deleted